refactor: extract post components
